requirements.txt
beautifulsoup4==4.6.0
boto==2.48.0
bs4==0.0.1
bz2file==0.98
certifi==2017.7.27.1
chardet==3.0.4
dill==0.2.7.1
gensim==3.0.1
idna==2.6
numpy==1.13.1
progressbar2==3.34.3
python-dateutil==2.6.1
python-utils==2.2.0
requests==2.18.4
Sastrawi==1.0.1
scikit-learn==0.19.0
scipy==0.19.1
six==1.11.0
smart-open==1.5.3
urllib3==1.22
